requirements.txt
PySide6>=6.6
yt-dlp>=2024.8.6
python-mpv>=1.0.7
pywin32>=306; sys_platform == "win32"
psutil>=5.9
